# specs/001-hungarian-learning-platform/contracts/api-specification.yaml
openapi: 3.0.3
info:
  title: Hungarian Language Learning Platform API
  description: |
    A comprehensive API for Korean speakers learning Hungarian, with special focus on sermon writing capabilities.

    **Features:**
    - CEFR-based progressive learning (A1-B2)
    - Real-time pronunciation assessment
    - Adaptive spaced repetition
    - Sermon composition tools
    - Offline synchronization

  version: 1.0.0
  contact:
    name: Hungarian Learning Platform Team
    email: dev@hungarian-learning.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.hungarian-learning.com/v1
    description: Production server
  - url: https://staging-api.hungarian-learning.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /progress:
    get:
      tags: [Progress]
      summary: Get user learning progress
      parameters:
        - name: cefr_level
          in: query
          schema:
            type: string
            enum: [A1, A2, B1, B2]
        - name: skill
          in: query
          schema:
            type: string
            enum: [reading, writing, speaking, listening]
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Progress data retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProgress'

    post:
      tags: [Progress]
      summary: Update learning progress
      description: Update user's progress after completing lessons or exercises
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressUpdate'
      responses:
        '200':
          description: Progress updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # Lesson Content
...
  /search:
    get:
      tags: [Search]
      summary: Search across content
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
        - name: type
          in: query
          description: Content type to search
          schema:
            type: string
            enum: [vocabulary, lessons, sermons, all]
            default: all
        - name: cefr_level
          in: query
          schema:
            type: string
            enum: [A1, A2, B1, B2]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SearchResults'

  # Offline Synchronization
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total_pages:
          type: integer
          minimum: 0
        total_items:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_previous:
          type: boolean

    # User Management Schemas
...
    UserProgress:
      type: object
      properties:
        overall_progress:
          type: object
          properties:
            current_cefr_level:
              type: string
              enum: [A1, A2, B1, B2]
            completion_percentage:
              type: number
              format: float
              minimum: 0
              maximum: 100
        skill_breakdown:
          type: object
          properties:
            reading:
              $ref: '#/components/schemas/SkillProgress'
            writing:
              $ref: '#/components/schemas/SkillProgress'
            speaking:
              $ref: '#/components/schemas/SkillProgress'
            listening:
              $ref: '#/components/schemas/SkillProgress'
        recent_activity:
          type: array
          items:
            $ref: '#/components/schemas/ActivitySummary'
        streak_info:
          type: object
          properties:
            current_streak:
              type: integer
            max_streak:
              type: integer
            last_study_date:
              type: string
              format: date
    SkillProgress:
      type: object
      properties:
        cefr_level:
          type: string
          enum: [A1, A2, B1, B2]
        progress_percentage:
          type: number
          format: float
          minimum: 0
          maximum: 100
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        last_assessment:
          type: string
          format: date-time
        strengths:
          type: array
          items:
            type: string
        areas_for_improvement:
          type: array
          items:
            type: string
    ActivitySummary:
      type: object
      properties:
        date:
          type: string
          format: date
        activity_type:
          type: string
          enum: [lesson, vocabulary_review, pronunciation_practice, sermon_writing]
        duration_minutes:
          type: integer
        items_completed:
          type: integer
        average_score:
          type: number
          format: float
    ProgressUpdate:
      type: object
      required:
        - lesson_id
        - completion_percentage
        - time_spent_minutes
      properties:
        lesson_id:
          type: string
          format: uuid
        completion_percentage:
          type: number
          format: float
          minimum: 0
          maximum: 100
        time_spent_minutes:
          type: integer
          minimum: 0
        exercise_scores:
          type: object
          additionalProperties:
            type: number
            format: float
        skill_assessments:
          type: object
          additionalProperties:
            type: number
            format: float

    # Lesson Content Schemas
...
    SearchResults:
      type: object
      properties:
        query:
          type: string
        total_results:
          type: integer
        search_time_ms:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        suggestions:
          type: array
          items:
            type: string
        pagination:
          $ref: '#/components/schemas/Pagination'
    SearchResult:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [vocabulary, lesson, sermon, audio]
        title:
          type: string
        content_preview:
          type: string
        cefr_level:
          type: string
        relevance_score:
          type: number
          format: float
        highlight:
          type: object
        url:
          type: string

    # Audio Schemas
...